test: 녹화 뷰 테스트를 view별 개별 파일로 분리, PresignPartView 테스트 추가, single_upload 테스트 제거
